fix: permalink parsing on timeline
The previous implementation was using DIR attr as a
selector, this can change based on region as it states
the direction of the language (ltr or rtl, etc). We are
now checking the link's href for instance of "status" this
is more future proof and will work no matter the user's region
or selected language.